python/pyproject.toml
[project]
name = "dotfiles"
version = "1.0"
description = "Prateek's dotfiles"
readme = "README.rst"
authors = [
    { name = "Prateek Raman", email = "vvnraman at gmail dot com" }
]
requires-python = ">=3.14"
dependencies = [
    "pydata-sphinx-theme>=0.16.1",
    "sphinx>=8.2.3",
    "sphinx-autobuild>=2025.8.25",
    "sphinx-design>=0.7.0",
    "sphinx-tabs>=3.4.7",
    "sphinxcontrib-mermaid>=1.0.0",
    "sphinxemoji>=0.3.1",
    "typer-slim>=0.20.0",
]

[project.scripts]
dotfiles = "dotfiles.main:app"

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"

[tool.basedpyright]
include = ["src", "../docs"]
exclude = [
  "**/__pycache__",
  "**/*.egg-info",
  "../docs/_build",
]
ignore = [
  "../home/",
]

reportMissingImports = "error"
reportImplicitStringConcatenation = false
reportUnannotatedClassAttribute  = false
reportImplicitOverride = false

[dependency-groups]
dev = [
    "basedpyright>=1.38.2",
    "black>=26.3.0",
    "isort>=8.0.1",
    "pytest>=8.4.2",
]
